Changed code for tests
