+++ dynamyc_list.sln
@@ -5,6 +5,8 @@ VisualStudioVersion = 16.0.29613.14
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "dynamyc_list", "dynamyc_list\dynamyc_list.vcxproj", "{6A379A7A-3A39-4A0D-9AD4-C7B72802616E}"
 EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "dynamic_listUnitTest", "dynamic_listUnitTest\dynamic_listUnitTest.vcxproj", "{77D2432E-07BA-4ACF-860D-187B386D8340}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|x64 = Debug|x64
@@ -21,6 +23,14 @@ Global
 		{6A379A7A-3A39-4A0D-9AD4-C7B72802616E}.Release|x64.Build.0 = Release|x64
 		{6A379A7A-3A39-4A0D-9AD4-C7B72802616E}.Release|x86.ActiveCfg = Release|Win32
 		{6A379A7A-3A39-4A0D-9AD4-C7B72802616E}.Release|x86.Build.0 = Release|Win32
+		{77D2432E-07BA-4ACF-860D-187B386D8340}.Debug|x64.ActiveCfg = Debug|x64
+		{77D2432E-07BA-4ACF-860D-187B386D8340}.Debug|x64.Build.0 = Debug|x64
+		{77D2432E-07BA-4ACF-860D-187B386D8340}.Debug|x86.ActiveCfg = Debug|Win32
+		{77D2432E-07BA-4ACF-860D-187B386D8340}.Debug|x86.Build.0 = Debug|Win32
+		{77D2432E-07BA-4ACF-860D-187B386D8340}.Release|x64.ActiveCfg = Release|x64
+		{77D2432E-07BA-4ACF-860D-187B386D8340}.Release|x64.Build.0 = Release|x64
+		{77D2432E-07BA-4ACF-860D-187B386D8340}.Release|x86.ActiveCfg = Release|Win32
+		{77D2432E-07BA-4ACF-860D-187B386D8340}.Release|x86.Build.0 = Release|Win32
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
